requirements.txt
python 3.4
Django 1.7.1
django-ckeditor 4.4.7
